Model/requirement.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
torch
transformers
chardet
PyPDF2
python-docx
google-generativeai
